+++ packages/fluentui/accessibility/src/behaviors/Pill/pillBehavior.ts
@@ -5,9 +5,9 @@ import { IS_FOCUSABLE_ATTRIBUTE } from '../../attributes';
 export const pillBehavior: Accessibility<PillBehaviorProps> = p => ({
   attributes: {
     root: {
-      role: p.actionable ? 'button' : 'none',
-      tabIndex: p.actionable ? 0 : -1,
-      [IS_FOCUSABLE_ATTRIBUTE]: p.actionable || p.role === 'option',
+      role: p.actionable || p.selectable ? 'button' : 'none',
+      tabIndex: p.actionable || p.selectable ? 0 : -1,
+      [IS_FOCUSABLE_ATTRIBUTE]: p.actionable || p.selectable || p.role === 'option',
       ...(p.selectable && {
         'aria-selected': p.selected,
       }),
@@ -15,10 +15,12 @@ export const pillBehavior: Accessibility<PillBehaviorProps> = p => ({
   },
   keyActions: {
     root: {
-      ...(p.actionable && {
+      ...(p.dismissible && {
         performDismiss: {
           keyCombinations: [{ keyCode: keyboardKey.Delete }, { keyCode: keyboardKey.Backspace }],
         },
+      }),
+      ...((p.selectable || p.actionable) && {
         performClick: {
           keyCombinations: [{ keyCode: keyboardKey.Enter }, { keyCode: SpacebarKey }],
         },
@@ -32,4 +34,5 @@ export type PillBehaviorProps = {
   selectable: boolean;
   selected: boolean;
   role: AriaRole;
+  dismissible: boolean;
 };