+++ packages/react-combobox/src/components/OptionGroup/OptionGroup.tsx
@@ -8,15 +8,11 @@ import type { ForwardRefComponent } from '@fluentui/react-utilities';
 /**
  * OptionGroup component: allows grouping of Option components within a Combobox
  */
-export const OptionGroup: ForwardRefComponent<OptionGroupProps> & { fluentComponentType?: string } = React.forwardRef(
-  (props, ref) => {
-    const state = useOptionGroup_unstable(props, ref);
+export const OptionGroup: ForwardRefComponent<OptionGroupProps> = React.forwardRef((props, ref) => {
+  const state = useOptionGroup_unstable(props, ref);
 
-    useOptionGroupStyles_unstable(state);
-    return renderOptionGroup_unstable(state);
-  },
-);
-
-OptionGroup.fluentComponentType = 'OptionGroup';
+  useOptionGroupStyles_unstable(state);
+  return renderOptionGroup_unstable(state);
+});
 
 OptionGroup.displayName = 'OptionGroup';