+++ packages/react-components/react-card/src/components/Card/useCard.ts
@@ -13,11 +13,19 @@ import { useFocusableGroup } from '@fluentui/react-tabster';
  * @param ref - reference to root HTMLElement of Card
  */
 export const useCard_unstable = (props: CardProps, ref: React.Ref<HTMLElement>): CardState => {
+  const { appearance = 'filled', focusMode = 'off' } = props;
+
+  const focusMap = {
+    off: undefined,
+    'no-tab': 'limitedTrapFocus',
+    'tab-exit': 'limited',
+    'tab-only': 'unlimited',
+  } as const;
+
   const groupperAttrs = useFocusableGroup({
-    tabBehavior: 'limitedTrapFocus',
+    tabBehavior: focusMap[focusMode],
   });
 
-  const { appearance = 'filled' } = props;
   return {
     appearance,
 
@@ -25,6 +33,7 @@ export const useCard_unstable = (props: CardProps, ref: React.Ref<HTMLElement>):
     root: getNativeElementProps(props.as || 'div', {
       ref,
       role: 'group',
+      tabIndex: focusMode !== 'off' ? 0 : undefined,
       ...groupperAttrs,
       ...props,
     }),